fix: correct calculation of spin observables

+++ calculation-helpers/concurrence-tests/combineSpinBasisValues.py
@@ -108,7 +108,7 @@ densityMatrixBuilderArray = np.array(
             ],
         ],
     ],
-    dtype=np.complex64,
+    dtype=np.complex128,
 )
