+++ calculation-helpers/concurrence-tests/spinToZBasisTransformation.py
@@ -24,7 +24,7 @@ def getMatrixInZBasis(
 ) -> np.ndarray:
     indexTranslationMap = {"uu": 0, "ud": 1, "du": 2, "dd": 3}
 
-    res = np.zeros((4, 4), dtype=np.complex64)
+    res = np.zeros((4, 4), dtype=np.complex128)
 
     for summandOfA in sva:
         plusA, has_iA, udketA, udbraA = summandOfA
